+++ src/main/java/org/openlmis/onenetwork/integration/web/StockEventController.java
@@ -1,0 +1,54 @@
+/*
+ * This program is part of the OpenLMIS logistics management information system platform software.
+ * Copyright © 2017 VillageReach
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms
+ * of the GNU Affero General Public License as published by the Free Software Foundation, either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Affero General Public License for more details. You should have received a copy of
+ * the GNU Affero General Public License along with this program. If not, see
+ * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org.
+ */
+
+package org.openlmis.onenetwork.integration.web;
+
+import org.openlmis.onenetwork.integration.configuration.IntegrationConfiguration;
+import org.openlmis.onenetwork.integration.dto.StockEvent;
+import org.openlmis.onenetwork.integration.service.StockEventBufferService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/integration")
+public class StockEventController {
+
+  private StockEventBufferService stockEventBufferService;
+  private IntegrationConfiguration integrationConfiguration;
+
+  @Autowired
+  public StockEventController(StockEventBufferService stockEventBufferService,
+                              IntegrationConfiguration integrationConfiguration) {
+    this.stockEventBufferService = stockEventBufferService;
+    this.integrationConfiguration = integrationConfiguration;
+  }
+
+  /**
+   * Receives stockEvent related data and send them to the buffer.
+   */
+  @PostMapping("/stockEvent")
+  @ResponseStatus(HttpStatus.OK)
+  public void sendStockEventToBuffer(
+      @RequestBody StockEvent stockEvent) {
+    if (integrationConfiguration.getEnable()) {
+      stockEventBufferService.add(stockEvent);
+    }
+  }
+}