+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
 New functionality added in a backwards-compatible manner:
+* [OLMIS-7305](https://openlmis.atlassian.net/browse/OLMIS-7305): Generating stock on hand csv file and sending to sftp server.
 * [OLMIS-7303](https://openlmis.atlassian.net/browse/OLMIS-7303): Generating facilities csv file and sending to sftp server.
 * [OLMIS-7304](https://openlmis.atlassian.net/browse/OLMIS-7304): Generating products csv file and sending to sftp server.
 * [OLMIS-7302](https://openlmis.atlassian.net/browse/OLMIS-7302): Created scaffold of one-network-integration-service.